+++ backend/profil-server/src/main/java/de/egladil/web/profil_server/cdi/AppLifecycleBean.java
@@ -1,0 +1,33 @@
+// =====================================================
+// Project: mk-gateway
+// (c) Heike Winkelvoß
+// =====================================================
+package de.egladil.web.profil_server.cdi;
+
+import javax.enterprise.context.ApplicationScoped;
+import javax.enterprise.event.Observes;
+
+import org.eclipse.microprofile.config.inject.ConfigProperty;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import io.quarkus.runtime.StartupEvent;
+
+/**
+ * AppLifecycleBean
+ */
+@ApplicationScoped
+public class AppLifecycleBean {
+
+	private static final Logger LOGGER = LoggerFactory.getLogger(AppLifecycleBean.class);
+
+	@ConfigProperty(name = "quarkus.http.cors.origins", defaultValue = "")
+	String corsAllowedOrigins;
+
+	void onStartup(@Observes final StartupEvent ev) {
+
+		LOGGER.info(" ===========>  quarkus.http.cors.origins={}", corsAllowedOrigins);
+
+	}
+
+}