+++ backend/profil-server/src/main/java/de/egladil/web/profil_server/filter/SecureHeadersFilter.java
@@ -34,8 +34,6 @@ public class SecureHeadersFilter implements ContainerResponseFilter {
 
 		final MultivaluedMap<String, Object> headers = responseContext.getHeaders();
 
-		addCORSHeaders(headers);
-
 		if (headers.get("Cache-Control") == null) {
 
 			headers.add("Cache-Control", "no-store, no-cache, must-revalidate, max-age=0");
@@ -88,45 +86,4 @@ public class SecureHeadersFilter implements ContainerResponseFilter {
 		}
 
 	}
-
-	/**
-	 * Theoretisch könnte man dies auch über die Quarkus-Konfigurationsprameter machen. Es hat sich aber herausgestellt, dass dies
-	 * zu
-	 * volatil ist und die Browser mit den Konstanten nicht gut zurecht kommen und CORS-Blockaden erzeugen. Daher bitte nicht in
-	 * application.properties mit den Quarkus-CORS-Parametern konfigurieren, sondern hier.
-	 *
-	 * @param headers
-	 */
-	private void addCORSHeaders(final MultivaluedMap<String, Object> headers) {
-
-		if (headers.get("Access-Control-Allow-Origin") == null) {
-
-			headers.add("Access-Control-Allow-Origin", config.getAllowedOrigin());
-		}
-
-		if (headers.get("Access-Control-Allow-Credentials") == null) {
-
-			headers.add("Access-Control-Allow-Credentials", "true");
-		}
-
-		// Achtung: mod-security verbietet standardmäßig PUT und DELETE.
-		// Daher parallel in /etc/apache2/sites-available/opa-wetterwachs.conf die rule 911100 für profil-server entfernen,
-		// sonst bekommt man 403
-		if (headers.get("Access-Control-Allow-Methods") == null) {
-
-			headers.add("Access-Control-Allow-Methods", "POST, PUT, GET, HEAD, OPTIONS, DELETE");
-		}
-
-		if (headers.get("Access-Control-Max-Age") == null) {
-
-			headers.add("Access-Control-Max-Age", "" + config.getAccessControlMaxAge());
-		}
-
-		if (headers.get("Access-Control-Allow-Headers") == null) {
-
-			headers.add("Access-Control-Allow-Headers",
-				"Content-Type, Accept, X-Requested-With, Authorization,X-SESSIONID,X-XSRF-TOKEN");
-		}
-	}
-
 }