+++ frontend/authprovider-ws/apps/profil-app/src/app/app.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { AuthService } from './services/auth.service';
 import { environment } from '../environments/environment';
 import { store } from './shared/store/app-data';
+import { VersionService } from './services/version.service';
 
 @Component({
 	// tslint:disable-next-line:component-selector
@@ -18,12 +19,13 @@ export class AppComponent implements OnInit {
 	api = environment.apiUrl;
 	logo = environment.assetsUrl + '/mja_logo.png';
 
-	constructor(private authService: AuthService
-		) { }
+	constructor(private authService: AuthService, public versionService: VersionService) { }
 
 	ngOnInit() {
 
 		store.updateBlockingIndicator(false);
+
+		this.versionService.ladeExpectedGuiVersion();
 
 		// nach dem redirect vom AuthProvider ist das die Stelle, an der die Anwendung wieder ankommt.
 		// Daher hier redirect-URL parsen