+++ frontend/authprovider-ws/build-env/profil-app/environment-a300.ts
@@ -5,7 +5,7 @@
 
 export const environment = {
   production: false,
-  version: '7.3.3',
+  version: '7.3.5',
   envName: 'a300',
   apiUrl: 'http://a300:9600/profil-api',
   assetsUrl: 'http://a300/profil-app/assets',