+++ RELEASE-NOTES.md
@@ -1,5 +1,11 @@
 # authprovider Release Notes
 
+__Release 7.3.5__:
+
+[Loading-Indicator bei diversen http-Requests](https://github.com/heike2718/authenticationprovider/issues/51)
+[Security upgrades](https://github.com/heike2718/authenticationprovider/issues/52)
+[Erwartete UI-Version vom Server holen](https://github.com/heike2718/authenticationprovider/issues/45)
+
 __Release 7.3.2__:
 
 [Kein Logout und keine Erfolgsmeldung nach dem Löschen des Benutzerkontos](https://github.com/heike2718/authenticationprovider/issues/44)